updated arrow to be within proper element
